+++ blogs/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
-from .models import Category, Blog, User
+from .models import Category, Blog
 
 
 admin.site.register(Category)
 admin.site.register(Blog)
-admin.site.register(User)
