+++ blogs/models.py
@@ -1,13 +1,5 @@
 from django.db import models
-from django.contrib.auth.models import AbstractUser
-
-
-class User(AbstractUser):
-    """A custom user model"""
-    image = models.ImageField(upload_to = 'papaya/images')
-    
-    def __str__(self):
-        return self.username
+from users.models import User
 
 
 class Category(models.Model):
