+++ templates/listing.html
@@ -1,9 +1,9 @@
-{% extends 'papaya/base.html' %}
+{% extends 'base.html' %}
 
 {% block content %}
 <div class="container well" id="bodyBox">
     {% for blog in blogs %}
-        <a href="{% url 'papaya:blogs_view' blog.id%}">
+        <a href="{% url 'blogs:blogs_view' blog.id%}">
             <div class="row blog">{{ blog }}
             </div>
         </a>
