+++ templates/profile_changepass.html
@@ -1,28 +1,28 @@
-{% extends 'papaya/base.html' %}
+{% extends 'base.html' %}
 
 {% block content %}
 <div class="container well" id="bodyBox">
     <div class="row">
-        <h1>Your profile</h1>
+        <h1>Change password:</h1>
     </div>
     <br>
     <div class="row">
         <form method="post" novalidate>
             {% csrf_token %}
-            <div class="form-group">
-                {{ form.username.label }}
-                {{ form.username.errors }}
-                {{ form.username }}
-            </div>
+            {{ error }}
             <div class="form-group">
                 {{ form.password.label }}
                 {{ form.password.errors }}
                 {{ form.password }}
             </div>
             <div class="form-group">
-                {{ form.password2.label }}
-                {{ form.password2.errors }}
-                {{ form.password2 }}
+                {{ form.password_new.label }}
+                {{ form.password_new.errors }}
+                {{ form.password_new }}
+            </div>
+            <div class="form-group">
+                {{ form.password_confirm.errors }}
+                {{ form.password_confirm }}
             </div>
             <br>
             <button class="btn btn-default" type="submit">Done</button>
