+++ templates/profile_login.html
@@ -1,4 +1,4 @@
-{% extends 'papaya/base.html' %}
+{% extends 'base.html' %}
 
 {% block content %}
 <div class="container well" id="bodyBox">
